+++ Travel.Solution/Travel.Service/LogicaNegocio/Interface/IAutorService.cs
@@ -7,10 +7,10 @@ namespace Travel.Core.LogicaNegocio.Interface
     public interface IAutorService
     {
         DataTable Autor_ObtAll();
-        DataTable Autor_ObtUno(int ID);
-        int Autor_Insertar(int ID, string Nombres, string Apellidos);
-        int Autor_Actualizar(int ID, string Nombres, string Apellidos);
+        DataTable Autor_ObtUno(double ID);
+        int Autor_Insertar(double ID, string Nombres, string Apellidos);
+        int Autor_Actualizar(double ID, string Nombres, string Apellidos);
         List<AutorDTO> AutorDTO_ObtAll();
-        AutorDTO AutorDTO_ObtUno(int ID);
+        AutorDTO AutorDTO_ObtUno(double ID);
     }
 }
